+++ CHANGES.rst
@@ -33,11 +33,15 @@ Unreleased
 - The C extension module now conforms to `PEP 489`_, closing `issue 1977`_.
   Thanks, `Adam Turner <pull 1978_>`_.
 
+- Fixed a "ValueError: min() arg is an empty sequence" error caused by strange
+  empty modules, found by `oss-fuzz`_.
+
 .. _issue 1746: https://github.com/nedbat/coveragepy/issues/1746
 .. _issue 1977: https://github.com/nedbat/coveragepy/issues/1977
 .. _pull 1978: https://github.com/nedbat/coveragepy/pull/1978
 .. _issue 1980: https://github.com/nedbat/coveragepy/issues/1980
 .. _PEP 489: https://peps.python.org/pep-0489
+.. _oss-fuzz: https://google.github.io/oss-fuzz/
 
 
 .. start-releases